requirements/base.txt
Django==2.1.5
six==1.12.0
sorl-thumbnail==12.5.0

